requirements.txt
fastapi
uvicorn
jinja2
python-multipart
selenium
beautifulsoup4
requests
urllib3
pandas
openpyxl
xlsxwriter
google-generativeai
pydantic
python-dotenv
aiofiles
python-jose
passlib
httpx
asyncpg
sqlalchemy
psycopg2-binary
psutil
undetected-chromedriver

# Hugging Face 한국어 모델 관련 패키지
torch
torchvision
torchaudio
transformers
accelerate
datasets
tokenizers
safetensors

# AI 에이전트 시스템 관련 패키지
google-generativeai
python-dotenv
dataclasses-json
